+++ lib/koolbase_flutter.dart
@@ -4,3 +4,5 @@ export 'src/koolbase.dart';
 export 'src/auth/auth_models.dart';
 export 'src/auth/auth_exceptions.dart';
 export 'src/auth/auth_client.dart';
+export 'src/storage/storage_client.dart';
+export 'src/storage/storage_models.dart';